Add index to audit event data
To improve the searchability of audit events, `supplierId` has been added
to draft brief audit events as a top level key in the audit event data.
To make this searchable in an effective way, this adds an index on it.

To do this, audit event data needed to be converted to jsonb. A partial
expression index is then added. Partial because it only indexes rows
that have the `supplierId` field in the data, and an expression index as
it only indexes the `supplierId` field instead of the whole data blob.