Allow data['supplierId'] to be int or string
The `data-supplier-id` flag passed to the view will always be a string,
however it's possible the supplierId in the data blob could be a number
of a string.

The change to the migration ensures that the index is built with the
string representation of whatever type is in the data blob.

The change to the `list_audits` view is to ensure that SQLAlchemy uses
the `->>` operator when creating the query for postgres. This means that
postgres knows that it can use the index we created, as they are both
specifically using the string version of the data.