Deprecate `supplier_id` from audit event data
We had one endpoint, `create_user`, which has been using the key
`supplier_id` instead of `supplierId` like we do everywhere else. This
updates that endpoint to fit with what we do elsewhere.

This means that we need to include that variation in the index. This was
achieved by using `COALESCE` in the expression used to optionally use
the `supplier_id` value if `supplierId` didn't exist. We also use
`COALESCE` in the where clause in the same way.

The `list_audits` endpoint also is updated to use coalesce. To make sure
that Postgres used the index we have to be really specific about using
the same expression on both sides of the query (in the index and in the
query itself).

This also adds a test to make sure that nothing untoward happens if an
audit event doesn't have supplier id in the data.