Update Docker base image to version v3.3.0
This commit updates the base image of this app to the latest version of
v3.3.0 of the Digital Marketplace docker image.

v3.3.0 is a rolling release that could theoretically be updated at any
time, however initially this app has been tied to a specific build.

Ticket: https://trello.com/c/ldIdPBzy