Reduce locking potential on DOS transition endpoint
Frameworks are now committed in a separate transaction, before the
briefs. This is to prevent the framework rows from being locked for any
longer than necessary. If there are lots of briefs to update, the
frameworks could be locked unnecessarily. The Framework rows are used
heavily by lots of api calls so that would be bad.

This also changes the way the briefs are loaded to only fetch the id and
framework_id columns. There's no need to pull all the data. This should
make the process faster.