Add suppliers 'active' flag
We want to be able to indicate in the database whether a supplier is
inactive (i.e. they have stopped existing, have merged with another
supplier). This commit adds a simple boolean flag to the supplier table.

We do this in two transactions so that we don't lock the entire table.

Co-authored-by: Ares Galamatis <argyris.galamatis@digital.cabinet-office.gov.uk>