Change test to be less specific
The error message caused the test to fail when changing indices to uniques. This was because the error message was hyper-specific. The only important part of the error message is that you can't double up DUNS numbers. Therefore I've made it slightly more generic to allow for the test to pass on either side of the migration.