DraftService: add lot_one_service_limit field & migration
this will eventually be used to mirror the value of the related
Lot.one_service_limit to allow it to be referenced in a unique constraint.

initially this has no constraints and is nullable - population of the field
will happen in a separate release to reduce the amount of time the
draft_services table needs to be locked. code is already modified to start
populating the field for new DraftServices.