Allow DMp team to create G12 drafts
For the G12 recovery, suppliers will only be able to submit the drafts that are part of the G12 recovery. We won't be allowing them to create any extra. This is already disabled in supplier-frontend, so we can remove it here as well.

Instead, we need to be able to create arbitrary G12 drafts. This is because some of the G12 recovery services don't have corresponding drafts, so we need to create them. And we don't want to add the suppliers to the list of G12 recovery suppliers until the recovery process opens, because it makes user-visible changes to the website.

This is a terrible hack, and I feel bad about introducing it to the codebase. However, we will be deleting all the G12 recovery code in a month's time, so I think it's OK.

I've tested this manually locally, and it seems to work just fine.