Add API endpoint to un-award a brief
We need the ability to un-award briefs. This will allow us to fix a brief that was accidentally awarded to the wrong supplier. This API method will not be exposed via any frontend app, so will be accessible to developers only.

Once a developer has un-awarded the brief using this new method, the buyer can then follow the usual process to award the brief to the correct supplier.

Update the validation rules on `awarded_at` accordingly.

Co-authored-by: Benjamin Gill <benjamin.gill@digital.cabinet-office.gov.uk>