Add 'deleted' service status
This status is used for when we want to remove a service page
completely and show a 404, rather than the `disabled` status which
leaves the page up with a warning message.

Alembic won't autodetect migrations like this (see [the docs](https://alembic.sqlalchemy.org/en/latest/autogenerate.html#what-does-autogenerate-detect-and-what-does-it-not-detect))
so I wrote the migration script myself. I have tested upgrades and
downgrades locally and it appears to be working, but I'dencourage
the reviewer to do the same.

https://trello.com/c/qqL3sqWV/2182-remove-supplier-92526-from-g-cloud