Make e-signature determination stable
Previously, we decided whether a framework supported esignature or not based on the date it went live. This worked. However, it means that if you need to make a non-esignature framework live now for testing, it breaks because it erroneously thinks it should support esignature.

Instead hard-code a list of the framework slugs that do not support esignature. This should be safer because all future frameworks will support esignature.