Make Employment Status question temporarily optional
Without this change we can't release the new question to production
without causing a period of time where buyers can't create new briefs.
This is because both the briefs FE and the API require the changes in
the other app to have already been deployed for everything to work.

With the question as optional in the API (but not in the briefs
frontend), we can release the API first, then the briefs FE, then
update the API to make the question required. This will allow us to
release everything without breaking things for our buyers.