Make employment status question required
In #1146 we added a new question to the brief creation journey. This
question is required, but we left it as optional to make the process of
releasing the question and updating the tests easier.

All that work's now done, so we can mark the question as required so
the validation schema in the API now matches what's generated from the
frameworks repo.