add error handling for social weighting
The social weighting question introduces a new validation case which was not previously handled. We can differenciate between the other 'oneOf' error handling as only one item is returned in the error path.  We then add some new logic to the _translate_json_schema_error to translate this validation error to the desired error code.

Co-authored-by: Alex Smith <alex.me.smith@gmail.com>