Add ability to filter for active users
It's very common in scripts that we only want to act on active users. So lots of scripts get a list of users, and then filter for the active users. It would save everyone time and effort if we did that filtering at the API instead.

Increase the page size (from 5 to 10) so all test results fit on one page.